whatsapp-bot/requirements.txt
flask==3.0.0
requests==2.31.0
openai==1.54.0
httpx==0.27.0
python-dotenv==1.0.0
supabase
gunicorn==21.2.0
